requirements.txt
# Runtime dependencies
croniter==2.0.7
PyYAML==6.0.2

# Test dependency
pytest==8.3.5
